Updated all user data int's to long's

+++ Code/SimpleAuthentication.Core/Providers/Facebook/MeResult.cs
@@ -9,7 +9,7 @@
         public string Link { get; set; }
         public string Username { get; set; }
         public string Email { get; set; }
-        public int Timezone { get; set; }
+        public long Timezone { get; set; }
         public string Locale { get; set; }
         public bool Verified { get; set; }
         public string UpdatedTime { get; set; }
